add eos loss

+++ fairseq/criterions/group_transformer_entropy.py
@@ -59,7 +59,7 @@ class GroupTransformerEntropy(FairseqCriterion):
         target = model.get_targets(sample, net_output).view(-1, 1)
         # add eos loss
         attns = net_output[1]['attn']
-        loss_eos = loss1 = attns[:,:,-1][:,:-1].sum()
+        loss_eos = loss1 = attns[:,:,-1][:,:-1].sum().half()
 
         non_pad_mask = target.ne(self.padding_idx)
         nll_loss = -lprobs.gather(dim=-1, index=target)[non_pad_mask]
